// src/enums/payment-status.enum.ts
export enum PaymentTransactionStatus {
  Pending = 'PENDING',
  Succeeded = 'SUCCEEDED',
  Failed = 'FAILED',
  Refunded = 'REFUNDED',
}

export enum PostPaymentStatus {
  NotRequired = 'NOT_REQUIRED',
  Pending = 'PENDING',
  Succeeded = 'SUCCEEDED',
  Failed = 'FAILED',
  Refunded = 'REFUNDED',
}